# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kommiuniti API",
    "description": "API for Kommiuniti App",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/registration/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Registrate a new user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/registration/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/registration/verify/{code}": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "Verfy a new account",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/community/getbyid/{id}": {
      "get": {
        "tags": [
          "Community"
        ],
        "description": "Get Community by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/community/getbycode/{code}": {
      "get": {
        "tags": [
          "Community"
        ],
        "description": "Find Community by the invite code",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/community/requests": {
      "get": {
        "tags": [
          "Community"
        ],
        "description": "Get all requests to a community, only admin can do this",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/community/requests/{communityid}": {
      "get": {
        "tags": [
          "Community"
        ],
        "description": "",
        "parameters": [
          {
            "name": "communityid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/community/acceptrequest": {
      "post": {
        "tags": [
          "Community"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/community/denierequest": {
      "post": {
        "tags": [
          "Community"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/community/create": {
      "post": {
        "tags": [
          "Community"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/debt/create": {
      "post": {
        "tags": [
          "Debt"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "amount": {
                  "example": "any"
                },
                "debitorId": {
                  "example": "any"
                },
                "creditorId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/debt/mine": {
      "get": {
        "tags": [
          "Debt"
        ],
        "description": "Endpoint to get all debts of a user.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/debt/balance": {
      "get": {
        "tags": [
          "Debt"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/debt/single/{id}": {
      "get": {
        "tags": [
          "Debt"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  }
}